Added tests and fixed some bugs
Bugs fixed
1. Invalid ansi sequences when both foreground and background colors are present
2. Crashes when rgb and fixed attributes encountered
3. Not throwing error for empty tags `<>`

+++ ziyy-core/src/common/position.rs
@@ -7,9 +7,9 @@ use std::{
 /// A position in source code.
 pub struct Position {
     /// Line in source code.
-    pub line: usize,
+    pub line: i32,
     /// Column in source code.
-    pub column: usize,
+    pub column: i32,
 }
 
 impl Default for Position {
@@ -20,13 +20,13 @@ impl Default for Position {
 
 impl Position {
     /// Creates a new Position.
-    pub fn new(line: usize, column: usize) -> Self {
+    pub fn new(line: i32, column: i32) -> Self {
         Self { line, column }
     }
 }
 
-impl AddAssign<(usize, usize)> for Position {
-    fn add_assign(&mut self, rhs: (usize, usize)) {
+impl AddAssign<(i32, i32)> for Position {
+    fn add_assign(&mut self, rhs: (i32, i32)) {
         let (line, column) = rhs;
         self.line += line;
         if line > 0 {
@@ -37,8 +37,8 @@ impl AddAssign<(usize, usize)> for Position {
     }
 }
 
-impl SubAssign<(usize, usize)> for Position {
-    fn sub_assign(&mut self, rhs: (usize, usize)) {
+impl SubAssign<(i32, i32)> for Position {
+    fn sub_assign(&mut self, rhs: (i32, i32)) {
         let (line, column) = rhs;
         self.line -= line;
         self.column -= column;
